Add react-diff-view library (#53)
Fixes #10 

This doesn't do much, but it does add a component that will allow us to embed a `Diff` component into a page/other component. I tested it by temporarily adding my new `DiffView` component into `App.tsx` and it rendered successfully (although it was admittedly very ugly to look at).

I imagine this component will be iterated on as we actually start using it, but I thought it worthwhile to get feedback on what I've done with TS and with the tests.

Note that the change to the README file was made by `prettier` on my system, so I've included those changes in this commit.