Set up Storybook (#65)
Fixes #4 

This just does a very basic set up of Storybook. It includes one example story which I took from the guide at https://storybook.js.org/basics/guide-react/. I imagine we'll remove that story as soon as we have our first "real" story. I also imagine we will enhance our instance of Storybook with addons, etc., but until we have an actual component for which to write a story I don't think it makes sense to do much more with it.

It works, in that you can start Storybook via `yarn storybook` and it doesn't break anything else about the app.

I wasn't sure if the deps for Storybook should go in `dependencies` or `devDependencies`, but I opted for the latter.