Access the user profile from the redux store (#136)
Fixes #97 

I'm having trouble getting the test for the logOut button working, to confirm that it dispatches `userActions.logOut`. I believe my code should be working, but jest reports that the mock was not called. I wonder if the problem is that I am using the `dispatch` method on the store object to do a real dispatch in `storeWithUser` and then after that trying to replace it with a mock via `jest.spyOn`? I need the original dispatch though, or the button won't appear.

Any advice anyone has would be appreciated.