Add ESLint reporter for jest (#141)
Fixes #140 

This is just an initial crack at getting this to work. I simply brought a couple of files over from addons-frontend, and added a new yarn command called `test-lint`, which runs the Jest tests and also the ESLint checks.

It's a bit hacky, but the only thing I could get to work, which I picked up from https://github.com/facebook/create-react-app/issues/2474#issuecomment-321837707.

The Jest reporter and utils files are still `js` files, as opposed to `tsx` files, and I guess we'll want to rewrite them as `tsx` files, as long as Jest allows that.

Just submitting this for initial thoughts and feedback.