Connect the FileTree component (#487)
~~Depends on #473.~~

It all works, but I'm not too sure about the tests. They feel a bit weird to me. Also, the Storybook story for `FileTree` is now broken, in that you cannot interact with it and expand the tree. I took a look but I'm not sure how to get it working.