Load a version into the store for the FileTree story (#503)
Fixes #502 

This doesn't actually fix #502, but it's a start. There was a problem before that we were not loading a version into the store, so when `toggleExpandedPath` was dispatched it generated an error. With this patch I can see that `toggleExpandedPath` is succeeding, and that `expandedPaths` in the store is being updated when we click on a folder, but the component isn't responding to that change. So it looks like the `FileTree` component in the story isn't actually connected to the store, even though we're using `renderWithStoreAndRouter`.

Do either of you have any ideas about why that might be?

As an aside, while doing some of this work on `FileTree` I keep running into situations in which it seems to me that it would make more sense to have `FileTree` get its `version` via `mapStateToProps` rather than being passed a `version` via `props` from its parent. I wonder if that would also fix this case as well, and if that's really what we should be doing here?