Create actions for selecting the next and previous file in the tree (#547)
Fixes #471 

This is just a proof of concept for executing this action which needs data from both the `fileTree` state and the `versions` state, and needs to modify the `versions` state, which I am doing by dispatching `versionActions.updateSelectedPath`. 

I thought I could get away with not having to access `getState` in the thunk, by accepting `currentPath` as a parameter, but we still need it to access the internal state of the `fileTree` reducer, because we also need to dispatch.

I may have gotten a number of things wrong here, because it still seems a bit convoluted to me. Submitting this for feedback on the approach.

@willdurand this is something I discussed with @kumar303 via Vidyo today, but please feel free to comment if you have any suggestions.