The worker seems to be... working!
At this point at least I know it's running, after commenting out the unused imports

Ok, so if we don't call `doTokenize` with `async`, then we get a promise back immediately, and then we can use postMessage in the worker to make something happen in the client, i.e., enable highlighting. But that will probably also force a re-render, which we don't want.

Still working, now need to deal with refractor

Definitely a problem with the refractor import - I tried converting everything to js, but didn't seem to help

Fixed the refractor issue, but still getting that "Error: Objects are not valid as a React child (found: [object Promise])"

It all seems to be working now, but trying to figure out how to wire it all together